refactor(Lotto): 당첨번호와 로또 클래스 분리
