+++ src/main/java/lotto/constant/LottoConstants.java
@@ -1,11 +1,9 @@
 package lotto.constant;
 
 public class LottoConstants {
-    public static final String INVALID_LOTTO_SIZE_MSG = "잘못된 로또 크기입니다.";
+    public static final String INVALID_LOTTO_SIZE_MSG = "로또 번호 개수가 잘못되었습니다.";
     public static final String DUPLICATE_NUMBER_EXIST_MSG = "로또 내부 중복된 값이 존재합니다.";
-    public static final String EMPTY_LOTTO_VALUE_MSG = "비어 있는 당첨 번호입니다.";
-    public static final String INVALID_LOTTO_INPUT_FORM = "잘못된 형식의 당첨 번호 입력입니다";
-    public static final String INVALID_RANGED_LOTTO_INPUT = "잘못된 범위의 당첨 번호가 포함된 입력입니다";
+    public static final String INVALID_RANGED_LOTTO_MSG = "잘못된 범위의 로또 번호가 존재합ㄴ디ㅏ";
 
 
     public static final int LOTTO_SIZE = 6;
