+++ src/main/java/lotto/constant/WinningLottoConstants.java
@@ -1,0 +1,9 @@
+package lotto.constant;
+
+public class WinningLottoConstants {
+    public static final String INVALID_WINNING_LOTTO_SIZE_MSG = "당첨 번호의 개수가 부족합니다.";
+    public static final String DUPLICATE_WINNING_NUMBER_EXIST_MSG = "당첨 번호 내부에 중복된 값이 존재합니다.";
+    public static final String INVALID_WINNING_LOTTO_INPUT_FORM_MSG = "잘못된 형식의 당첨 번호 입력입니다";
+    public static final String INVALID_RANGED_WINNING_LOTTO_INPUT_MSG = "잘못된 범위의 당첨 번호가 포함된 입력입니다";
+
+}
